Updated job spec and code cleanup for lendi
